Crate: add individual value parsing for padding/margin/inset/border
