+++ takumi/src/style/properties/mod.rs
@@ -25,18 +25,14 @@ pub use sides::*;
 use cosmic_text::{Align, FamilyOwned, Weight};
 use cssparser::{BasicParseError, ParseError, Parser};
 use image::imageops::FilterType;
-use merge::{Merge, option::overwrite_none};
 use serde::{Deserialize, Serialize};
 use taffy::{
-  CompactLength, GridTemplateRepetition, MaxTrackSizingFunction, MinTrackSizingFunction, Size,
-  Style as TaffyStyle, TrackSizingFunction,
+  CompactLength, GridTemplateRepetition, MaxTrackSizingFunction, MinTrackSizingFunction,
+  TrackSizingFunction,
 };
 use ts_rs::TS;
 
-use crate::{
-  core::{DEFAULT_FONT_SIZE, DEFAULT_LINE_HEIGHT_SCALER, viewport::RenderContext},
-  impl_from_taffy_enum,
-};
+use crate::{core::viewport::RenderContext, impl_from_taffy_enum};
 
 type ParseResult<'i, T, E = BasicParseError<'i>> = Result<T, ParseError<'i, E>>;
 
@@ -323,162 +319,6 @@ impl From<FontFamily> for FamilyOwned {
   }
 }
 
-/// Represents the resolved font style for a text node.
-///
-/// This struct contains the resolved font style properties after inheriting
-/// from parent elements and converting relative units to absolute values.
-#[derive(Debug, Clone)]
-pub struct FontStyle {
-  /// Font size in pixels for text rendering
-  pub font_size: f32,
-  /// Line height as an absolute value in pixels
-  pub line_height: f32,
-  /// Font weight for text rendering
-  pub font_weight: Weight,
-  /// Maximum number of lines for text before truncation
-  pub line_clamp: Option<u32>,
-  /// Font family name for text rendering
-  pub font_family: Option<FamilyOwned>,
-  /// Letter spacing for text rendering in em units (relative to font size)
-  pub letter_spacing: Option<f32>,
-  /// Text alignment within the element
-  pub text_align: Option<Align>,
-  /// How text should be overflowed
-  pub text_overflow: TextOverflow,
-  /// Text color for child text elements
-  pub color: LinearGradientOrColor,
-}
-
-/// Main styling structure that contains all layout and visual properties.
-///
-/// This structure combines both layout properties (like width, height, padding)
-/// and inheritable properties (like font settings, colors) that can be passed
-/// down to child elements.
-#[derive(Debug, Clone, Deserialize, Serialize, TS)]
-#[serde(default)]
-#[ts(export, optional_fields)]
-pub struct Style {
-  /// Display algorithm to use for the element
-  pub display: Display,
-  /// Width of the element
-  pub width: LengthUnit,
-  /// Height of the element
-  pub height: LengthUnit,
-  /// Max width of the element
-  pub max_width: LengthUnit,
-  /// Max height of the element
-  pub max_height: LengthUnit,
-  /// Min width of the element
-  pub min_width: LengthUnit,
-  /// Min height of the element
-  pub min_height: LengthUnit,
-  /// Aspect ratio of the element (width/height)
-  pub aspect_ratio: Option<f32>,
-  /// Internal spacing around the element's content (top, right, bottom, left)
-  pub padding: Sides<LengthUnit>,
-  /// External spacing around the element (top, right, bottom, left)
-  pub margin: Sides<LengthUnit>,
-  /// Positioning offsets (top, right, bottom, left) from the element's normal position
-  pub inset: Sides<LengthUnit>,
-  /// Direction of flex layout (row or column)
-  pub flex_direction: FlexDirection,
-  /// How a single grid item is aligned along the inline (row) axis, overriding the container's justify-items value
-  pub justify_self: Option<AlignItems>,
-  /// How items are aligned along the main axis
-  pub justify_content: Option<JustifyContent>,
-  /// How lines are aligned within the flex container when there's extra space in the cross axis
-  pub align_content: Option<JustifyContent>,
-  /// How grid items are aligned along the inline (row) axis within their grid areas
-  pub justify_items: Option<AlignItems>,
-  /// How items are aligned along the cross axis
-  pub align_items: Option<AlignItems>,
-  /// How a single item is aligned along the cross axis, overriding the container's align-items value
-  pub align_self: Option<AlignItems>,
-  /// How flex items should wrap
-  pub flex_wrap: FlexWrap,
-  /// The initial main size of the flex item before growing or shrinking
-  pub flex_basis: LengthUnit,
-  /// Positioning method (relative, absolute, etc.)
-  pub position: Position,
-  /// Spacing between rows and columns in flex or grid layouts
-  pub gap: Gap,
-  /// How much the flex item should grow relative to other flex items when positive free space is distributed
-  pub flex_grow: f32,
-  /// How much the flex item should shrink relative to other flex items when negative free space is distributed
-  pub flex_shrink: f32,
-  /// Width of the element's border on each side (top, right, bottom, left)
-  pub border_width: Sides<LengthUnit>,
-  /// How images should be fitted within their container
-  pub object_fit: ObjectFit,
-  /// Background gradient(s)
-  pub background_image: Option<LinearGradients>,
-  /// Background color for the element
-  pub background_color: Option<Color>,
-  /// Box shadow effect for the element
-  pub box_shadow: Option<BoxShadows>,
-  /// Controls the size of implicitly-created grid columns
-  pub grid_auto_columns: Option<Vec<GridTrackSize>>,
-  /// Controls the size of implicitly-created grid rows
-  pub grid_auto_rows: Option<Vec<GridTrackSize>>,
-  /// Controls how auto-placed items are inserted in the grid
-  pub grid_auto_flow: Option<GridAutoFlow>,
-  /// Specifies a grid item's size and location within the grid column
-  pub grid_column: Option<GridLine>,
-  /// Specifies a grid item's size and location within the grid row
-  pub grid_row: Option<GridLine>,
-  /// Defines the line names and track sizing functions of the grid columns
-  pub grid_template_columns: Option<Vec<GridTemplateComponent>>,
-  /// Defines the line names and track sizing functions of the grid rows
-  pub grid_template_rows: Option<Vec<GridTemplateComponent>>,
-  /// Inheritable style properties that cascade to child elements
-  #[serde(flatten)]
-  pub inheritable_style: InheritableStyle,
-}
-
-impl Default for Style {
-  fn default() -> Self {
-    Self {
-      display: Display::Flex,
-      margin: Sides([LengthUnit::zero(); 4]),
-      padding: Sides([LengthUnit::zero(); 4]),
-      width: Default::default(),
-      height: Default::default(),
-      max_width: Default::default(),
-      max_height: Default::default(),
-      min_width: Default::default(),
-      min_height: Default::default(),
-      aspect_ratio: None,
-      inset: Default::default(),
-      flex_direction: Default::default(),
-      justify_content: Default::default(),
-      align_content: Default::default(),
-      justify_self: Default::default(),
-      justify_items: Default::default(),
-      align_items: Default::default(),
-      align_self: Default::default(),
-      position: Default::default(),
-      gap: Default::default(),
-      flex_grow: 0.0,
-      flex_shrink: 1.0,
-      flex_basis: Default::default(),
-      flex_wrap: FlexWrap::NoWrap,
-      border_width: Sides([LengthUnit::zero(); 4]),
-      object_fit: Default::default(),
-      box_shadow: Default::default(),
-      background_color: None,
-      background_image: None,
-      grid_auto_columns: None,
-      grid_auto_rows: None,
-      grid_auto_flow: None,
-      grid_column: None,
-      grid_row: None,
-      grid_template_columns: None,
-      grid_template_rows: None,
-      inheritable_style: Default::default(),
-    }
-  }
-}
-
 /// Represents a grid track sizing function with serde support
 #[derive(Debug, Clone, Deserialize, Serialize, TS, PartialEq)]
 #[serde(rename_all = "kebab-case")]
@@ -646,7 +486,7 @@ impl GridTemplateComponent {
   /// # Returns
   ///
   /// A `taffy::GridTemplateComponent` that can be used with the Taffy layout engine
-  fn to_taffy(&self, context: &RenderContext) -> taffy::GridTemplateComponent<String> {
+  pub fn to_taffy(&self, context: &RenderContext) -> taffy::GridTemplateComponent<String> {
     match self {
       Self::Single(track_size) => {
         taffy::GridTemplateComponent::Single(track_size.to_min_max(context))
@@ -718,165 +558,4 @@ impl From<ImageScalingAlgorithm> for FilterType {
   }
 }
 
-/// Style properties that can be inherited by child elements.
-///
-/// These properties are typically passed down from parent to child elements
-/// in the layout hierarchy, such as font settings and colors.
-#[derive(Debug, Clone, Deserialize, Serialize, Default, Merge, TS)]
-#[merge(strategy = overwrite_none)]
-#[ts(optional_fields, export)]
-pub struct InheritableStyle {
-  /// How text should be overflowed
-  pub text_overflow: Option<TextOverflow>,
-  /// Color of the element's border
-  pub border_color: Option<Color>,
-  /// Text color for child text elements
-  pub color: Option<LinearGradientOrColor>,
-  /// Font size in pixels for text rendering
-  pub font_size: Option<LengthUnit>,
-  /// Font family name for text rendering
-  pub font_family: Option<FontFamily>,
-  /// Line height multiplier for text spacing
-  pub line_height: Option<LengthUnit>,
-  /// Font weight for text rendering
-  pub font_weight: Option<FontWeight>,
-  /// Maximum number of lines for text before truncation
-  pub line_clamp: Option<u32>,
-  /// Corner radius for rounded borders
-  pub border_radius: Option<Sides<LengthUnit>>,
-  /// Text alignment within the element
-  pub text_align: Option<TextAlign>,
-  /// Additional spacing between characters in text
-  /// Positive values increase spacing, negative values decrease spacing
-  pub letter_spacing: Option<LengthUnit>,
-  /// Controls how images are scaled when rendered
-  /// This property determines the algorithm used for image scaling
-  pub image_rendering: Option<ImageScalingAlgorithm>,
-}
-
-impl Style {
-  /// Resolves the style to a `TaffyStyle`.
-  /// Converts this style to a Taffy-compatible style for layout calculations.
-  ///
-  /// # Arguments
-  ///
-  /// * `context` - The render context containing viewport information for unit resolution
-  ///
-  /// # Returns
-  ///
-  /// A `TaffyStyle` that can be used with the Taffy layout engine
-  pub fn resolve_to_taffy_style(&self, context: &RenderContext) -> TaffyStyle {
-    TaffyStyle {
-      size: Size {
-        width: self.width.resolve_to_dimension(context),
-        height: self.height.resolve_to_dimension(context),
-      },
-      border: resolve_length_unit_rect_to_length_percentage(context, self.border_width.into()),
-      padding: resolve_length_unit_rect_to_length_percentage(context, self.padding.into()),
-      inset: resolve_length_unit_rect_to_length_percentage_auto(context, self.inset.into()),
-      margin: resolve_length_unit_rect_to_length_percentage_auto(context, self.margin.into()),
-      display: self.display.into(),
-      flex_direction: self.flex_direction.into(),
-      position: self.position.into(),
-      justify_content: self.justify_content.map(Into::into),
-      align_content: self.align_content.map(Into::into),
-      justify_items: self.justify_items.map(Into::into),
-      flex_grow: self.flex_grow,
-      align_items: self.align_items.map(Into::into),
-      gap: self.gap.resolve_to_size(context),
-      flex_basis: self.flex_basis.resolve_to_dimension(context),
-      flex_shrink: self.flex_shrink,
-      flex_wrap: self.flex_wrap.into(),
-      min_size: Size {
-        width: self.min_width.resolve_to_dimension(context),
-        height: self.min_height.resolve_to_dimension(context),
-      },
-      max_size: Size {
-        width: self.max_width.resolve_to_dimension(context),
-        height: self.max_height.resolve_to_dimension(context),
-      },
-      // Convert grid properties to Taffy types using only public API
-      grid_auto_columns: self.grid_auto_columns.as_ref().map_or_else(Vec::new, |v| {
-        v.iter().map(|s| s.to_min_max(context)).collect()
-      }),
-      grid_auto_rows: self.grid_auto_rows.as_ref().map_or_else(Vec::new, |v| {
-        v.iter().map(|s| s.to_min_max(context)).collect()
-      }),
-      grid_auto_flow: self.grid_auto_flow.unwrap_or_default().into(),
-      grid_column: self
-        .grid_column
-        .as_ref()
-        .map_or_else(Default::default, |line| line.clone().into()),
-      grid_row: self
-        .grid_row
-        .as_ref()
-        .map_or_else(Default::default, |line| line.clone().into()),
-      grid_template_columns: self
-        .grid_template_columns
-        .as_ref()
-        .map_or_else(Vec::new, |v| {
-          v.iter().map(|s| s.to_taffy(context)).collect()
-        }),
-      grid_template_rows: self.grid_template_rows.as_ref().map_or_else(Vec::new, |v| {
-        v.iter().map(|s| s.to_taffy(context)).collect()
-      }),
-      aspect_ratio: self.aspect_ratio,
-      align_self: self.align_self.map(Into::into),
-      justify_self: self.justify_self.map(Into::into),
-      ..Default::default()
-    }
-  }
-
-  /// Resolves the style to a `ResolvedFontStyle`.
-  /// Resolves inheritable style properties to concrete values for text rendering.
-  ///
-  /// This method combines the element's inheritable styles with default values
-  /// to produce a complete font style specification.
-  ///
-  /// # Arguments
-  ///
-  /// * `context` - The render context containing viewport information for unit resolution
-  ///
-  /// # Returns
-  ///
-  /// A `ResolvedFontStyle` with all font-related properties resolved to concrete values
-  pub fn resolve_to_font_style(&self, context: &RenderContext) -> FontStyle {
-    let font_size = self
-      .inheritable_style
-      .font_size
-      .map(|f| f.resolve_to_px(context))
-      .unwrap_or(DEFAULT_FONT_SIZE);
-
-    let line_height = self
-      .inheritable_style
-      .line_height
-      .map(|f| f.resolve_to_px(context))
-      .unwrap_or_else(|| font_size * DEFAULT_LINE_HEIGHT_SCALER);
-
-    FontStyle {
-      color: self
-        .inheritable_style
-        .color
-        .clone()
-        .unwrap_or(LinearGradientOrColor::Color(Color([0, 0, 0, 255]))),
-      font_size,
-      line_height,
-      font_weight: self
-        .inheritable_style
-        .font_weight
-        .unwrap_or_default()
-        .into(),
-      line_clamp: self.inheritable_style.line_clamp,
-      font_family: self.inheritable_style.font_family.clone().map(Into::into),
-      letter_spacing: self
-        .inheritable_style
-        .letter_spacing
-        .map(|spacing| spacing.resolve_to_px(context) / context.parent_font_size),
-      text_align: self.inheritable_style.text_align.and_then(Into::into),
-      text_overflow: self
-        .inheritable_style
-        .text_overflow
-        .unwrap_or(TextOverflow::Clip),
-    }
-  }
-}
+// Style, InheritableStyle, and FontStyle moved to `style/stylesheets.rs`
